Finish the general part conversion for type fix

+++ general/base/CommonDatapack.cpp
@@ -145,8 +145,8 @@ void CommonDatapack::parseEvents()
 void CommonDatapack::parseMonsters()
 {
     #ifndef CATCHCHALLENGER_CLASS_MASTER
-    if(CommonSettingsServer::commonSettingsServer.rates_xp<=0 || CommonSettingsServer::commonSettingsServer.rates_xp==1.0 ||
-            CommonSettingsServer::commonSettingsServer.rates_xp_pow<=0 || CommonSettingsServer::commonSettingsServer.rates_xp_pow==1.0)
+    if(CommonSettingsServer::commonSettingsServer.rates_xp<=0 || static_cast<double>(CommonSettingsServer::commonSettingsServer.rates_xp)==1.0 ||
+            CommonSettingsServer::commonSettingsServer.rates_xp_pow<=0 || static_cast<double>(CommonSettingsServer::commonSettingsServer.rates_xp_pow)==1.0)
     {
         monsterRateApplied=false;
         CommonSettingsServer::commonSettingsServer.rates_xp=1.0;
